Put back closure methods and doc to OpenAIProtocol
